Cast dispatched expert ids to int32

+++ python/sglang/srt/eplb/expert_location_dispatch.py
@@ -80,9 +80,9 @@ def topk_ids_logical_to_physical(
         return topk_ids
 
     if info.ep_dispatch_algorithm == "static":
-        return _topk_ids_logical_to_physical_static(topk_ids, info)
+        return _topk_ids_logical_to_physical_static(topk_ids, info).to(torch.int32)
     if info.ep_dispatch_algorithm in ["dynamic", "fake"]:
-        return _topk_ids_logical_to_physical_dynamic(topk_ids, info)
+        return _topk_ids_logical_to_physical_dynamic(topk_ids, info).to(torch.int32)
     raise NotImplementedError(f"Unknown algorithm {info.ep_dispatch_algorithm}")
